local_module/requirements.txt
pandas==2.1.4
openpyxl==3.1.2
pillow==10.2.0
torch==2.10.0
scikit-learn==1.3.2
